Prototype 1: Added Special Attack
Upon start of the game the player has no special attack selected.
Special Abilities can be added to the drop table just like any other loot. a new child so from drop called dropAbility was added. to create a new dropAbility use the asset menu to create a new dropability. Give it a name and add a spell description to the so.
Make sure to also add this new ability to the drop table.