Prototype2: Fire Trap
Added a new type of SpellDescription called RuneSpellDescription. The Spell places a FireTrap prefab at the position of the player. Values for the FireTrap can be edited in the prefab.
If an enemy enters the trap a fireshot will spawn. The trap deals damage to all enemies within the explosion range defined in the prefab.
A FireTrap is setup in the testing scene. To test the ability immediately add FireTrap to Eve_Player in the special attack spellslot.